test: triggers scale with KEDA
